c/partition_allocator: moved partition allocator to cluster/scheduling
Moved `cluster::partition_allocator` to `cluster/scheduling` source
directory to separate allocation/scheduling abstractions from rest of
the cluster module implementation.
