+++ src/v/cluster/members_backend.h
@@ -3,7 +3,7 @@
 #include "cluster/fwd.h"
 #include "cluster/members_manager.h"
 #include "cluster/members_table.h"
-#include "cluster/partition_allocator.h"
+#include "cluster/scheduling/partition_allocator.h"
 #include "cluster/types.h"
 #include "model/metadata.h"
 #include "raft/consensus.h"
