+++ src/v/cluster/members_manager.cc
@@ -14,7 +14,7 @@
 #include "cluster/fwd.h"
 #include "cluster/logger.h"
 #include "cluster/members_table.h"
-#include "cluster/partition_allocator.h"
+#include "cluster/scheduling/partition_allocator.h"
 #include "cluster/types.h"
 #include "config/configuration.h"
 #include "model/metadata.h"
