+++ src/v/cluster/scheduling/partition_allocator.cc
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "cluster/partition_allocator.h"
+#include "cluster/scheduling/partition_allocator.h"
 
 #include "cluster/logger.h"
 #include "model/metadata.h"
