+++ src/v/cluster/scheduling/partition_allocator.h
