+++ src/v/cluster/security_frontend.cc
@@ -16,7 +16,6 @@
 #include "cluster/controller_stm.h"
 #include "cluster/errc.h"
 #include "cluster/logger.h"
-#include "cluster/partition_allocator.h"
 #include "cluster/partition_leaders_table.h"
 #include "cluster/types.h"
 #include "model/errc.h"
