+++ src/v/cluster/topics_frontend.cc
@@ -14,7 +14,7 @@
 #include "cluster/controller_stm.h"
 #include "cluster/errc.h"
 #include "cluster/logger.h"
-#include "cluster/partition_allocator.h"
+#include "cluster/scheduling/partition_allocator.h"
 #include "cluster/partition_leaders_table.h"
 #include "cluster/types.h"
 #include "model/errc.h"
